hwval(feature) Add Redfish URI validations
Initial set of Redfish URI validations. Checks to see if certain Redfish
URIs contain the proper fields and are of the expected type. Added more
output types like Info and Warning. Reorganized the code to be more
modular. Added ability to use IP addresses. Improved test selection to
allow for an entire group to be selected.